auth methods moved to auth, legacy methods from httpauth have been removed
